+++ apscheduler/schedulers/base.py
@@ -19,9 +19,10 @@ from apscheduler.job import Job
 from apscheduler.triggers.base import BaseTrigger
 from apscheduler.util import asbool, asint, astimezone, maybe_ref, timedelta_seconds, undefined
 from apscheduler.events import (
-    SchedulerEvent, JobEvent, EVENT_SCHEDULER_START, EVENT_SCHEDULER_SHUTDOWN,
+    SchedulerEvent, JobEvent, JobSubmissionEvent, EVENT_SCHEDULER_START, EVENT_SCHEDULER_SHUTDOWN,
     EVENT_JOBSTORE_ADDED, EVENT_JOBSTORE_REMOVED, EVENT_ALL, EVENT_JOB_MODIFIED, EVENT_JOB_REMOVED,
-    EVENT_JOB_ADDED, EVENT_EXECUTOR_ADDED, EVENT_EXECUTOR_REMOVED, EVENT_ALL_JOBS_REMOVED)
+    EVENT_JOB_ADDED, EVENT_EXECUTOR_ADDED, EVENT_EXECUTOR_REMOVED, EVENT_ALL_JOBS_REMOVED,
+    EVENT_JOB_SUBMITTED, EVENT_JOB_MAX_INSTANCES)
 
 
 class BaseScheduler(six.with_metaclass(ABCMeta)):
@@ -859,6 +860,7 @@ class BaseScheduler(six.with_metaclass(ABCMeta)):
         self._logger.debug('Looking for jobs to run')
         now = datetime.now(self.timezone)
         next_wakeup_time = None
+        events = []
 
         with self._jobstores_lock:
             for jobstore_alias, jobstore in six.iteritems(self._jobstores):
@@ -882,12 +884,19 @@ class BaseScheduler(six.with_metaclass(ABCMeta)):
                             self._logger.warning(
                                 'Execution of job "%s" skipped: maximum number of running '
                                 'instances reached (%d)', job, job.max_instances)
+                            event = JobSubmissionEvent(EVENT_JOB_MAX_INSTANCES, job.id,
+                                                       jobstore_alias, run_times)
+                            events.append(event)
                         except:
                             self._logger.exception('Error submitting job "%s" to executor "%s"',
                                                    job, job.executor)
+                        else:
+                            event = JobSubmissionEvent(EVENT_JOB_SUBMITTED, job.id, jobstore_alias,
+                                                       run_times)
+                            events.append(event)
 
-                        # Update the job if it has a next execution
-                        # time. Otherwise remove it from the job store.
+                        # Update the job if it has a next execution time.
+                        # Otherwise remove it from the job store.
                         job_next_run = job.trigger.get_next_fire_time(run_times[-1], now)
                         if job_next_run:
                             job._modify(next_run_time=job_next_run)
@@ -901,6 +910,10 @@ class BaseScheduler(six.with_metaclass(ABCMeta)):
                 if jobstore_next_run_time and (next_wakeup_time is None or
                                                jobstore_next_run_time < next_wakeup_time):
                     next_wakeup_time = jobstore_next_run_time.astimezone(self.timezone)
+
+        # Dispatch collected events
+        for event in events:
+            self._dispatch_event(event)
 
         # Determine the delay until this method should be called again
         if next_wakeup_time is not None: